+++ CHANGELOG.md
@@ -33,7 +33,7 @@ Emojis for the following are chosen based on [gitmoji](https://gitmoji.dev/).
 - The messages that tell the user a noun is already plural have been translated to the keyboard's language.
 - The keyboard height has been increased for landscape mode on iPads.
 - Characters on keys have been made larger so they reflect the system keyboards better.
-<!--- All letter, number and special keys now pop up after being pressed. -->
+- All letter, number and special keys now pop up after being pressed. 
 <!--- Hold to select characters have been redesigned to reflect the addition of keys popping up. -->
 <!--- All App Store media has been redone to reflect these changes.-->